重写了所有的messagebox(除了登录注册界面)

+++ mainwindow.cpp
@@ -400,7 +400,11 @@ void MainWindow::on_OpenFileBtn_clicked()
 
     file_full = QFileDialog::getOpenFileName(this,"Open File",QDir::currentPath());//打开文件选择
     if (file_full.isEmpty()){
-        QMessageBox::information(this,"Erro Message","请选择文件");
+        MsgBox *msgbox = new MsgBox(3,QStringLiteral("请选择文件"));
+        msgbox->exec();
+//        QMessageBox message(QMessageBox::NoIcon,"Erro Message","请选择文件");
+//        message.setIconPixmap(QPixmap(":/new/mainwindow/pictures/system_waring.png"));
+//        message.exec();
         fileOpenFlag = false;
 
     }else{
@@ -503,7 +507,8 @@ void MainWindow::handleResults(int value)
         EncryptionItem *v1 = ui->MidStaWidget->findChild<EncryptionItem*>(f_progressBar->objectName());
         v1->encryptStaBtn->clicked();
 
-        QMessageBox::information(NULL,tr("成功"),tr("加密完成！"),QMessageBox::Yes,NULL);
+        MsgBox *msgbox = new MsgBox(4,QStringLiteral("分享成功！"));
+        msgbox->exec();
 ///////////////////////////////删除加密完成的项目
 //        EncryptionItem *v1 = ui->MidStaWidget->findChild<EncryptionItem*>(f_progressBar->objectName());
 //        //qDebug()<<name;
@@ -562,7 +567,11 @@ void MainWindow::startEncryptThread(){
 
 //批量删除按钮
 void MainWindow::on_pushButton_3_clicked()
+
 {
+    MsgBox *msgbox = new MsgBox(1,QStringLiteral("确认删除这些已加密条目吗？"));
+    int nRes = msgbox->exec();
+    if (nRes == QDialog::Accepted){
     QSqlQuery query(db);
     bool success = query.exec("select * from Decryption where oemp_id='"+User_ID+"'");
     if(!success){
@@ -590,7 +599,6 @@ void MainWindow::on_pushButton_3_clicked()
                           }
 
                       }
-
                   }
                   else{  //若没找到则继续查找
                       continue;
@@ -605,6 +613,9 @@ void MainWindow::on_pushButton_3_clicked()
               QVBoxLayout *newVbox = new QVBoxLayout();
               newVbox->addWidget(newScrollArea);
               decryptionViewController->setLayout(newVbox);
+    }
+    }else{
+
     }
 
 }
@@ -651,10 +662,12 @@ void MainWindow::getFileID(){
                           m1->downloadBtn->setText("申请中");
                           m1->fileDescription->setText("正在等待对方应答，请等待！");
                           disconnect(m1->downloadBtn,SIGNAL(clicked(bool)),this,SLOT(getFileID()));
-                           QMessageBox::warning(this,tr("Success"),tr("申请成功请等待！"),QMessageBox::Yes);
+                          MsgBox *msgbox = new MsgBox(4,QStringLiteral("申请成功，请等待！"));
+                          msgbox->exec();
                       }
                       else{
-                           QMessageBox::warning(this,tr("error"),tr("申请失败!"),QMessageBox::Yes);
+                          MsgBox *msgbox = new MsgBox(2,QStringLiteral("申请失败！"));
+                          msgbox->exec();
                       }
                   }
               }
@@ -715,7 +728,8 @@ void MainWindow::OssDownLoadFile(){
                     disconnect(m1->downloadBtn,SIGNAL(clicked(bool)),this,SLOT(OssDownLoadFile()));//删除原有信号槽
                 }
                 else{
-                    QMessageBox::warning(this,tr("error"),tr("下载失败"),QMessageBox::Yes);
+                    MsgBox *msgbox = new MsgBox(2,QStringLiteral("下载失败！"));
+                    msgbox->exec();
                 }
             }
         }
@@ -840,7 +854,8 @@ void MainWindow::addFriendToDatabase(QString name){
     else{
         while (query.next()) {
             if(query.record().value("friend_nickname").toString()==name){
-                QMessageBox::warning(this,tr("Warning"),tr("已经是好友！"),QMessageBox::Yes);
+                MsgBox *msgbox = new MsgBox(3,QStringLiteral("已经是你的好友了！"));
+                msgbox->exec();
                 return;
             }
             else{
@@ -854,13 +869,15 @@ void MainWindow::addFriendToDatabase(QString name){
     qDebug()<<name;
     if(success4){
         if(query.size()==0){
-            QMessageBox::warning(this,tr("Failed!"),tr("好友不存在"),QMessageBox::Yes);
+            MsgBox *msgbox = new MsgBox(2,QStringLiteral("好友不存在"));
+            msgbox->exec();
             return;
         }
         else{
             while (query.next()) {
                 if(query.record().value("emp_id").toString().isEmpty()){
-                    QMessageBox::warning(this,tr("Failed!"),tr("好友不存在"),QMessageBox::Yes);
+                    MsgBox *msgbox = new MsgBox(2,QStringLiteral("好友不存在"));
+                    msgbox->exec();
                     return;
                 }
                 else{
@@ -874,7 +891,8 @@ void MainWindow::addFriendToDatabase(QString name){
     }
     else {
         qDebug()<<"查询失败";
-        QMessageBox::warning(this,tr("Failed!"),tr("好友不存在"),QMessageBox::Yes);
+        MsgBox *msgbox = new MsgBox(2,QStringLiteral("好友不存在"));
+        msgbox->exec();
         return;
     }
     bool success1 = query.exec("select *from employee where emp_id = '"+User_ID+"'");
@@ -901,10 +919,12 @@ void MainWindow::addFriendToDatabase(QString name){
         int i = friendListWidget->count();
         i++;
         friendListWidget->insertItem(i,name);
-         QMessageBox::warning(this,tr("Success"),tr("添加好友成功！"),QMessageBox::Yes);
+        MsgBox *msgbox = new MsgBox(4,QStringLiteral("添加好友成功！"));
+        msgbox->exec();
     }
     else{
-        QMessageBox::warning(this,tr("Failed!"),tr("添加好友失败！"),QMessageBox::Yes);
+        MsgBox *msgbox = new MsgBox(2,QStringLiteral("添加好友失败！"));
+        msgbox->exec();
     }
 }
 
@@ -918,7 +938,8 @@ void MainWindow::inforDlgaddFriend(QString name){
     else{
         while (query.next()) {
             if(query.record().value("friend_nickname").toString()==name){
-                QMessageBox::warning(this,tr("Warning"),tr("已经是好友！"),QMessageBox::Yes);
+                MsgBox *msgbox = new MsgBox(2,QStringLiteral("已经是你的好友了"));
+                msgbox->exec();
                 return;
             }
             else{
@@ -932,13 +953,15 @@ void MainWindow::inforDlgaddFriend(QString name){
     qDebug()<<name;
     if(success4){
         if(query.size()==0){
-            QMessageBox::warning(this,tr("Failed!"),tr("好友不存在"),QMessageBox::Yes);
+            MsgBox *msgbox = new MsgBox(2,QStringLiteral("好友不存在！"));
+            msgbox->exec();
             return;
         }
         else{
             while (query.next()) {
                 if(query.record().value("emp_id").toString().isEmpty()){
-                    QMessageBox::warning(this,tr("Failed!"),tr("好友不存在"),QMessageBox::Yes);
+                    MsgBox *msgbox = new MsgBox(2,QStringLiteral("好友不存在！"));
+                    msgbox->exec();
                     return;
                 }
                 else{
@@ -952,7 +975,8 @@ void MainWindow::inforDlgaddFriend(QString name){
     }
     else {
         qDebug()<<"查询失败";
-        QMessageBox::warning(this,tr("Failed!"),tr("好友不存在"),QMessageBox::Yes);
+        MsgBox *msgbox = new MsgBox(2,QStringLiteral("好友不存在!"));
+        msgbox->exec();
         return;
     }
     bool success1 = query.exec("select *from employee where emp_id = '"+User_ID+"'");
@@ -979,10 +1003,12 @@ void MainWindow::inforDlgaddFriend(QString name){
         int i = friendListWidget->count();
         i++;
         friendListWidget->insertItem(i,name);
-         QMessageBox::warning(this,tr("Success"),tr("添加好友成功！"),QMessageBox::Yes);
+        MsgBox *msgbox = new MsgBox(4,QStringLiteral("添加好友成功！"));
+        msgbox->exec();
     }
     else{
-        QMessageBox::warning(this,tr("Failed!"),tr("添加好友失败！"),QMessageBox::Yes);
+        MsgBox *msgbox = new MsgBox(2,QStringLiteral("添加好友失败！"));
+        msgbox->exec();
     }
 }
 
@@ -1015,7 +1041,8 @@ void MainWindow::on_pushButton_4_clicked()
     if(originalText.startsWith("{",Qt::CaseSensitive)){
         LinkInsert(originalText);
         qDebug()<<originalText;
-        QMessageBox::information(this,tr("Success"),tr("已读取链接信息"),QMessageBox::Yes);
+        MsgBox *msgbox = new MsgBox(4,QStringLiteral("已读取链接信息！"));
+        msgbox->exec();
     }else{
     linkDialog = new DelinkDialog();
     connect(linkDialog,SIGNAL(sendLinkToMain(QString)),this,SLOT(LinkInsert(QString)));
@@ -1128,6 +1155,9 @@ void MainWindow::on_pushButton_8_clicked()
 //已加密文件批量删除按钮
 void MainWindow::on_pushButton_5_clicked()
 {
+    MsgBox *msgbox = new MsgBox(1,QStringLiteral("确认删除这些已加密条目吗？"));
+    int nRes = msgbox->exec();
+    if (nRes == QDialog::Accepted){
     int flag = 0;
     QString file_id_d;
         QSqlQuery query3(db);
@@ -1156,11 +1186,16 @@ void MainWindow::on_pushButton_5_clicked()
          }
         on_pushButton_8_clicked();
         if(flag == 1){
-            QMessageBox::information(NULL, "success", "成功批量删除条目！");
+            MsgBox *msgbox = new MsgBox(4,QStringLiteral("成功批量删除条目！"));
+            msgbox->exec();
         }else if(flag == 0){
-            QMessageBox::warning(NULL, "warning", "请选择需要删除的条目！");
+            MsgBox *msgbox = new MsgBox(2,QStringLiteral("请选择需要删除的条目"));
+            msgbox->exec();
         }
         flag = 0;
+    }else{
+
+    }
 }
 //批量分享按钮响应
 void MainWindow::on_pushButton_6_clicked()
@@ -1190,52 +1225,49 @@ void MainWindow::on_pushButton_6_clicked()
             grpDlg = new groupSendDialog();
             grpDlg->show();
         }else if(flag == 0){
-            QMessageBox::warning(NULL, "warning", "请选择需要批量分享的条目！");
+            MsgBox *msgbox = new MsgBox(2,QStringLiteral("请选择需要批量分享的条目！"));
+            msgbox->exec();
         }
         flag = 0;    
 }
 
 //已加密文件单独删除按钮
 void MainWindow::on_deleteBtn_clicked(){
-    QMessageBox::StandardButton reply;
-        reply = QMessageBox::question(this, tr("确认删除"),
-                                        "确认删除这一已加密条目吗？",
-                                        QMessageBox::Yes | QMessageBox::No);
-        if (reply == QMessageBox::Yes){
-    QPushButton* button = qobject_cast<QPushButton*>(sender());
-    QString name = button->objectName();
+    MsgBox *msgbox = new MsgBox(1,QStringLiteral("确认删除这一已加密条目吗？"));
+    int nRes = msgbox->exec();
+    if (nRes == QDialog::Accepted){
+        QPushButton* button = qobject_cast<QPushButton*>(sender());
+        QString name = button->objectName();
 
 
-    FinishEncryptionItem *f2 = ui->MidStaWidget->findChild<FinishEncryptionItem*>(name);
-    delete f2;
-    qDebug()<<name;
+        FinishEncryptionItem *f2 = ui->MidStaWidget->findChild<FinishEncryptionItem*>(name);
+        delete f2;
+        qDebug()<<name;
 
-    QSqlQuery query(db);
-    bool success = query.exec("delete from varticle where article_id = '"+name+"'");
-    if(success){
-        QMessageBox::information(NULL, "success", "成功删除已加密文件条目！");
-    }
-
-    delete finishViewController->layout();
-    QWidget *newItemWidget = new QWidget();
-    newItemWidget->setLayout(finishViewController->vbox);
-    finScrollArea->setStyleSheet("border:0;padding:0;spacing:0;");
-    finScrollArea->setWidget(newItemWidget);
-    QVBoxLayout *newVbox = new QVBoxLayout();
-    newVbox->addWidget(finScrollArea);
-    finishViewController->setLayout(newVbox);
-    }
-        else if (reply == QMessageBox::No){
+        QSqlQuery query(db);
+        bool success = query.exec("delete from varticle where article_id = '"+name+"'");
+        if(success){
+            MsgBox *msgbox = new MsgBox(4,QStringLiteral("成功删除已加密条目！"));
+            msgbox->exec();
         }
 
+        delete finishViewController->layout();
+        QWidget *newItemWidget = new QWidget();
+        newItemWidget->setLayout(finishViewController->vbox);
+        finScrollArea->setStyleSheet("border:0;padding:0;spacing:0;");
+        finScrollArea->setWidget(newItemWidget);
+        QVBoxLayout *newVbox = new QVBoxLayout();
+        newVbox->addWidget(finScrollArea);
+        finishViewController->setLayout(newVbox);
+        }
+    else if (nRes == QDialog::Rejected){
+    }
 }
 //已解密文件单独删除按钮
 void MainWindow::on_deleteBtn2_clicked(){
-    QMessageBox::StandardButton reply;
-        reply = QMessageBox::question(this, tr("确认删除"),
-                                        "确认删除这一已解密条目吗？",
-                                        QMessageBox::Yes | QMessageBox::No);
-        if (reply == QMessageBox::Yes){
+    MsgBox *msgbox = new MsgBox(1,QStringLiteral("确认删除这一已解密条目吗？"));
+    int nRes = msgbox->exec();
+    if (nRes == QDialog::Accepted){
             QPushButton* button = qobject_cast<QPushButton*>(sender());
             QString name = button->objectName();
 
@@ -1246,8 +1278,8 @@ void MainWindow::on_deleteBtn2_clicked(){
             QSqlQuery query(db);
             bool success = query.exec("delete from Decryption where file_id = '"+name+"'");
             if(success){
-                QMessageBox::information(NULL, "success", "成功删除已解密文件条目！");
-                RequestNum--;
+                MsgBox *msgbox = new MsgBox(4,QStringLiteral("成功删除已解密条目！"));
+                msgbox->exec();
             }
 
             delete finishViewController2->layout();
@@ -1260,7 +1292,7 @@ void MainWindow::on_deleteBtn2_clicked(){
             newVbox->addWidget(finScrollArea);
             finishViewController2->setLayout(newVbox);
         }
-        else if (reply == QMessageBox::No){
+        else if (nRes == QDialog::Rejected){
         }
 }
 
@@ -1455,7 +1487,8 @@ void MainWindow::FileIsAllowed(){
                      QVBoxLayout *newVbox = new QVBoxLayout();
                      newVbox->addWidget(newScrollArea);
                      decryptionViewController->setLayout(newVbox);
-                     QMessageBox::warning(this,tr("Success"),tr("文件解密同意并成功解密"),QMessageBox::Yes);
+                     MsgBox *msgbox = new MsgBox(4,QStringLiteral("文件解密同意并成功解密"));
+                     msgbox->exec();
                  }
                  else{
                      return;
@@ -1514,6 +1547,9 @@ void MainWindow::NumDown(){
 //已解密文件批量删除按钮
 void MainWindow::on_pushButton_7_clicked()
 {
+    MsgBox *msgbox = new MsgBox(1,QStringLiteral("确认删除这些已解密条目吗？"));
+    int nRes = msgbox->exec();
+    if (nRes == QDialog::Accepted){
     int flag = 0;
     QString file_id;
         QSqlQuery query3(db);
@@ -1542,11 +1578,16 @@ void MainWindow::on_pushButton_7_clicked()
          }
         on_pushButton_9_clicked();
         if(flag == 1){
-            QMessageBox::information(NULL, "success", "成功批量删除条目！");
+            MsgBox *msgbox = new MsgBox(4,QStringLiteral("成功批量删除条目！"));
+            msgbox->exec();
         }else if(flag == 0){
-            QMessageBox::warning(NULL, "warning", "请选择需要删除的条目！");
+            MsgBox *msgbox = new MsgBox(2,QStringLiteral("请选择需要删除的条目！"));
+            msgbox->exec();
         }
         flag = 0;
+    }else{
+
+    }
 }
 
 //已解密页面全选按钮
@@ -1651,7 +1692,8 @@ void MainWindow::LinkInsert(QString link){
         }
     }
     if(Link_empid == NULL){
-        QMessageBox::warning(this,tr("ERROR"),tr("无效链接！"),QMessageBox::Yes);
+        MsgBox *msgbox = new MsgBox(2,QStringLiteral("无效链接！"));
+        msgbox->exec();
         return;
     }
     QUuid strid = QUuid::createUuid();
@@ -1691,7 +1733,8 @@ void MainWindow::LinkInsert(QString link){
 }
 //信号槽：接收到解密线程发送的解密失败信号
 void MainWindow::RecDecryptionFailed(){
-    QMessageBox::warning(this,tr("error"),tr("文件解密失败"),QMessageBox::Yes);
+    MsgBox *msgbox = new MsgBox(2,QStringLiteral("文件解密失败"));
+    msgbox->exec();
     decryptionFlag=1;
 }
 
@@ -1722,7 +1765,8 @@ void MainWindow::on_pushButton_12_clicked()
         QImage img;        //加载图像
         if(!(img.load(m_fileName)))
         {
-            QMessageBox::information(this,tr("打开图像失败"),tr("打开图像失败!"));
+            MsgBox *msgbox = new MsgBox(2,QStringLiteral("打开图像失败！"));
+            msgbox->exec();
             return;
         }
 
@@ -1731,9 +1775,11 @@ void MainWindow::on_pushButton_12_clicked()
         QString originalText = strQRCode.section("&&",1,1);
         if(originalText.startsWith("{",Qt::CaseSensitive)){
             LinkInsert(originalText);
-            QMessageBox::information(this,tr("Success"),tr("已读取二维码信息"),QMessageBox::Yes);
+            MsgBox *msgbox = new MsgBox(4,QStringLiteral("已读取二维码信息！"));
+            msgbox->exec();
         }else{
-            QMessageBox::warning(this,tr("fail"),tr("二维码读取失败"),QMessageBox::Yes);
+            MsgBox *msgbox = new MsgBox(2,QStringLiteral("二维码读取失败！"));
+            msgbox->exec();
         }
     }
 
@@ -1814,11 +1860,9 @@ void MainWindow::on_pushButton_13_clicked()
 
 void MainWindow::ShowNewDownDialog(QString id){
     qDebug()<<"shoudao";
-    QMessageBox::StandardButton reply;
-        reply = QMessageBox::question(this, tr("有新消息"),
-                                        "确定下载这条新消息？",
-                                        QMessageBox::Yes | QMessageBox::No);
-        if(reply == QMessageBox::Yes){
+    MsgBox *msgbox = new MsgBox(1,QStringLiteral("确认下载这一待解密文件吗？"));
+    int reply = msgbox->exec();
+        if(reply == QDialog::Accepted){
             //跳转到解密页面，开始下载
             DecryptionItem *m1 = ui->MidStaWidget->findChild<DecryptionItem *>(id+"decryption");
             if(m1==NULL){
@@ -1829,12 +1873,14 @@ void MainWindow::ShowNewDownDialog(QString id){
                 on_DecryptionBtn_clicked();
             }
         }
-        else if(reply ==QMessageBox::No){
+        else if(reply == QDialog::Rejected){
 
         }
 }
 void MainWindow::closeEvent(QCloseEvent *event){
-    if(QMessageBox::warning(this,"退出","确认退出吗？",QMessageBox::Yes | QMessageBox::No)==QMessageBox::Yes){
+    MsgBox *msgbox = new MsgBox(1,QStringLiteral("确认关闭系统吗？"));
+    int reply = msgbox->exec();
+    if(reply == QDialog::Accepted){
         event->accept();
     }else{
         event->ignore();
